+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * refactor: rename the term 'large' to 'basic'
 * cleanup: Basic classes do not require soft and hard limits
 * cleanup: orderBy should be non-empty-array
+* feat: initial version of basic repository
 
 ## 0.3.0