+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * feat: initial version of basic repository
 * refactor: consolidate repeated orderBy logic to `OrderByUtil`
 * refactor: use `configure()` method for subclasses to configure the repository
+* test: rearrange test directories
 
 ## 0.3.0